backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
python-multipart==0.0.17

python-dotenv==1.0.1
pypdf==5.1.0
requests==2.31.0
streamlit==1.35.0
groq==0.37.1